Add worker network tag to worker nodes
This fixes the issue where we exposed a service and the load balancer
was allocated an external IP, but it kept timing out trying to reach it.
A firewall rule was being created for the worker tag, but the node
didn't have the worker tag so now it is added during the Insert call.